refactor: page.clientを作成
